Created datastructure for model checking
Made a copy of the datastructure of the LTL on-the-fly algorithm and the mem-efficient verification algorithm. The copies can be modified to be used for the model-cheker while the old datastructure remains untouched for reference purposes.

In this copy:  Renamed TemporalProperties to AtomicProposition, this corresponds better with logic terminoligy.
Added AtomicPropositions to the model StateNode. Those properties hold in that state and can be used while model checking.

Created static variables as an alphabet for model checking to make comparison of the desired state in the ltl nodes with the real state in the model nodes more easy.